separated home and landing page
